requirements.txt
# Core Framework
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0

# RAG
chromadb>=0.5.0
pymupdf>=1.25.0
rank-bm25>=0.2.2
jieba>=0.42.1

# Web Search
duckduckgo-search>=7.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0

# API & Testing
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pytest-asyncio>=0.23.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
